cs-job-create: Permit creation of `synth' runvars
This will be useful for some hostalloc_* runvars which we are going to
introduce shortly.

This is going to be the way to set a runvar which is not copied by
cs-bisection-step or cs-adjust-flight.  Using `synth' for this is
arguably slightly wrong but it does the right thing in all existing
cases.  The alternative would be a schema change.

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>
Acked-by: Ian Campbell <ian.campbell@citrix.com>
---
v2: Defer some of the discussion to later commit messages.