ts-windows-install: Bump RAM from 512M to 1536M
Sadly we lack a proper infrastructure for handing guest memory size
and fitting it into the host.  We don't currently even have in our
database the amount of RAM the host has.

The recommended minimum for Windows is 2Gby.  But we have (in the
Cambridge instance) some hosts which only have 4Gby.  We allow 512 for
dom0.  We need to leave some spare for Xen and other overheads.  So
set this to 1536 (1.5Gby) which should fit even during localhost
migration.

Later we may want to do something more sophisticated, like adjusting
this to 2048 if the host has enough, or only running the test on
bigger hosts, or something.

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>
CC: Wei Liu <wei.liu2@citrix.com>
CC: Andrew Cooper <andrew.cooper3@citrix.com>
Acked-by: Ian Campbell <ian.campbell@citrix.com>
---
v2: 1536 not 1700.
    Mention migration in the commit message.