Debian HVM guests: Honour guest disk and ram size runvars
Allow runvars to specify guest disk and ram size, turning previous
values into defaults:

The default disk size for the guest is `10000M' which is not going to
be sufficient for nested HVM tests.  We are going to want to use a
larger disk size for the nested L1.  The appropriate disk_size will be
defined in make-flight, in forthcoming changes.

Also, also allow ram size to be defined by runvar.  The runvar takes
precedence over the default (which is calculated based on host RAM
etc.)

No functional change with current sets of runvars.

Signed-off-by: Robert Ho <robert.hu@intel.com>
Acked-by: Ian Jackson <ian.jackson@eu.citrix.com>
Signed-off-by: Ian Jackson <ian.jackson@eu.citrix.com>