ap-push: Use refs/heads/ for destinations
When the destination is a branch, specify refs/heads/ explicitly.
This makes ap-push work even if the ref does not yet exist on the
destination.

There is no functional change for an existing installation pushing to
an existing branch.  But for a hypothetical new installation, this
would be necessary.

And, more relevantly, when new "branches" are invented, the use of an
existing ap-push case as a template will generate a new case which
creates the branch as is necessary.

I leave the more complex osstest case alone.  It's not clear to me
whether the destination ref not existing is an installation problem of
such severity that indeed ap-push should fail.

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>
Acked-by: Ian Campbell <ian.campbell@citrix.com>