standalone-generate-dump-flight-runvars: memoise ap-fetch
This makes it _much_ faster.

Also you can say AP_FETCH_MEMO_KEEP=1 to make it reuse old
information, which is useful for making comparisons.

For a further speed improvement, one can use `eatmydata'.  This is not
the default because it risks corruption of `standalone.db' which is
used for other purposes too.  Add a comment about possibly improving
this.

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>
Acked-by: Ian Campbell <ian.campbell@citrix.com>